+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 1.17.3
 
 - changed: GP (drawing) object visibility icon show when there is a viewport-rendering conflict
+- changed: Better display of camera properties in list
 - added: info-hint in drawing's displays toggle channel
 
 1.17.2